+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+- Add support for BIP-86 taproot receive addresses
 - Show coin subtotals in 'My portfolio'
-- Transactio details: make transaction ID copyable without opening the block explorer
+- Transaction details: make transaction ID copyable without opening the block explorer
 
 ## 4.31.1 [2022-02-07]
 - Bundle BitBox02 Multi firmware version v9.9.1